requirements.txt
torch
numpy
python-chess
gymnasium
tqdm
